ทดลองนะ

+++ src/main/java/com/springoeb/employee/controller/EmployeeController.java
@@ -17,9 +17,10 @@ import java.io.IOException;
 import java.sql.Date;
 import java.sql.Time;
 import java.text.ParseException;
-import java.time.LocalTime;
 import java.util.ArrayList;
+import java.util.HashSet;
 import java.util.List;
+import java.util.Set;
 
 /**
  * Created by bighead on 6/9/17.
@@ -38,20 +39,26 @@ public class EmployeeController {
     @Autowired
     private WorkHistoryService workHistoryService;
 
+    private final String EMP_PATH = "/WEB-INF/employee/";
+
     @RequestMapping("/index")
-    public String toEmployeeIndex(Model model, HttpSession session) {
-        int branchNo = (Integer) (session.getAttribute("branchno"));
-        List<Employee> employees = employeeService.findByBranchNo(branchNo);
-        model.addAttribute("employees", employees);
-        return "/WEB-INF/empindex.jsp";
+    public String toEmployeeIndex() {
+        return EMP_PATH + "empindex.jsp";
     }
 
     @PostMapping("/manageemployee")
-    public void addAndEditEmployee(@ModelAttribute("employee") Employee employee, HttpSession session, HttpServletRequest request, HttpServletResponse response) throws IOException {
+    public void addAndEditEmployee(@ModelAttribute("employee") Employee employee, HttpSession session, HttpServletRequest request, HttpServletResponse response) throws Exception {
         int branchNo = (Integer) (session.getAttribute("branchno"));
         employee.setBranchNo(branchNo);
-        employeeService.save(employee);
-        response.sendRedirect(request.getContextPath() + "/employee/index");
+        if (!employeeService.chkDuplicateEmpName(employee.getEmpName(), branchNo)) {
+            employeeService.save(employee);
+        } else {
+            if (employee.getEmpNo() != null && employeeService.findByEmpNoAndBranchNo(employee.getEmpNo(), branchNo).getEmpName().equals(employee.getEmpName())) {
+                employeeService.save(employee);
+            } else {
+                throw new Exception();
+            }
+        }
     }
 
     @Transactional
@@ -59,7 +66,6 @@ public class EmployeeController {
     public void delEmployee(@PathVariable("empNo") int empNo, HttpSession session, HttpServletRequest request, HttpServletResponse response) throws IOException {
         int branchNo = (Integer) (session.getAttribute("branchno"));
         employeeService.removeByEmpNoAndBranchNo(empNo, branchNo);
-        response.sendRedirect(request.getContextPath() + "/employee/index");
     }
 
     @GetMapping("/ajax/getemployee/{empNo}")
@@ -72,25 +78,41 @@ public class EmployeeController {
         return json;
     }
 
+    @PostMapping("/ajax/getemployees")
+    @ResponseBody
+    public String getJsonEmployees(HttpSession session) throws JsonProcessingException {
+        int branchNo = (Integer) (session.getAttribute("branchno"));
+        List<Employee> employee = employeeService.findByBranchNo(branchNo);
+        ObjectMapper mapper = new ObjectMapper();
+        String json = mapper.writeValueAsString(employee);
+        return json;
+    }
+
     //----------------------------------------------------------------------------------------------------------//
 
     @RequestMapping("/position")
-    public String toEmployeePositionIndex(Model model) {
-        List<EmployeePosition> employeePositions = employeePositionService.findAll();
-        model.addAttribute("employeePositions", employeePositions);
-        return "/WEB-INF/emppos.jsp";
+    public String toEmployeePositionIndex() {
+        return EMP_PATH+"emppos.jsp";
     }
 
     @PostMapping("/manageemployeeposition")
-    public void addAndEditPosition(@ModelAttribute("employeePosition") EmployeePosition employeePosition, HttpServletRequest request, HttpServletResponse response) throws IOException {
-        employeePositionService.save(employeePosition);
-        response.sendRedirect(request.getContextPath() + "/employee/position");
+    public void addAndEditPosition(@ModelAttribute("employeePosition") EmployeePosition employeePosition, HttpServletRequest request, HttpServletResponse response, HttpSession session) throws Exception {
+        int branchNo = (Integer) (session.getAttribute("branchno"));
+        if (!employeePositionService.chkDuplicateEmpPosisitonName(employeePosition.getEmpPosName(), branchNo)) {
+            employeePositionService.save(employeePosition);
+        } else {
+            throw new Exception();
+        }
     }
 
     @Transactional
     @GetMapping("/deleteemployeeposition/{empPosNo}")
     public void removeByEmpPosNo(@PathVariable("empPosNo") int empPosNo, HttpServletRequest request, HttpServletResponse response) throws IOException {
-        employeePositionService.removeByEmpPosNo(empPosNo);
+        if (employeeTableService.canDeleteEmpPos(empPosNo)) {
+            employeePositionService.removeByEmpPosNo(empPosNo);
+        } else {
+            employeePositionService.unused(employeePositionService.findByEmpPosNo(empPosNo));
+        }
         response.sendRedirect(request.getContextPath() + "/employee/position");
     }
 
@@ -103,16 +125,23 @@ public class EmployeeController {
         return json;
     }
 
+    @PostMapping("/ajax/getemployeepositions")
+    @ResponseBody
+    public String getJsonEmployeePositions(HttpSession session) throws JsonProcessingException {
+        List<EmployeePosition> employeePositions = employeePositionService.findAll();
+        ObjectMapper mapper = new ObjectMapper();
+        String json = mapper.writeValueAsString(employeePositions);
+        return json;
+    }
+
     //----------------------------------------------------------------------------------------------------------//
 
     @RequestMapping("/workhistory")
     public String toEmployeeWorkHistory(Model model, HttpSession session) {
         int branchNo = (Integer) (session.getAttribute("branchno"));
-        List<WorkHistory> workHistories = workHistoryService.findAll(branchNo);
         List<Employee> employees = employeeService.findByBranchNo(branchNo);
-        model.addAttribute("workHistories", workHistories);
         model.addAttribute("employees", employees);
-        return "/WEB-INF/empworkhist.jsp";
+        return EMP_PATH+"empworkhist.jsp";
     }
 
     @PostMapping("/manageworkhistory")
@@ -123,7 +152,8 @@ public class EmployeeController {
         String stEmpNos[] = request.getParameterValues("empNo");
         String stWorkHours[] = request.getParameterValues("workHour");
         String stWorkMinutes[] = request.getParameterValues("workMin");
-
+        String stWorkHistNo = request.getParameter("workHistNo");
+        Integer workHistNo = stWorkHistNo == null ? null : Integer.parseInt(stWorkHistNo);
         Date workDate = Date.valueOf(stWorkDate);
 
         for (int i = 0; i < stEmpNos.length; i++) {
@@ -133,7 +163,10 @@ public class EmployeeController {
             String payType = employee.getPayType();
             double pay = employee.getPay();
             int workHour = Integer.parseInt(stWorkHours[i]);
-            int workMinute = Integer.parseInt(stWorkMinutes[i]);
+            int workMinute = 0;
+            if (stWorkMinutes[i] != null && !stWorkMinutes[i].equals("")) {
+                workMinute = Integer.parseInt(stWorkMinutes[i]);
+            }
             double sumWorkHour = workHour + (workMinute * 5 / 300.0);
             if (payType.equals(Employee.HOUR)) {
                 pay = pay * sumWorkHour;
@@ -144,6 +177,9 @@ public class EmployeeController {
             workHistory.setWorkMin(workMinute);
             workHistory.setWorkDate(workDate);
             workHistory.setWorkPay(pay);
+            if(workHistNo != null){
+                workHistory.setWorkHistNo(workHistNo);
+            }
             workHistories.add(workHistory);
         }
 
@@ -161,6 +197,68 @@ public class EmployeeController {
         response.sendRedirect(request.getContextPath() + "/employee/workhistory");
     }
 
+    @PostMapping("/ajax/getworkhistories")
+    @ResponseBody
+    public String getJsonWorkHistories(HttpSession session) throws JsonProcessingException {
+        int branchNo = (Integer) (session.getAttribute("branchno"));
+        List<WorkHistory> workHistories = workHistoryService.findAll(branchNo);
+        for (WorkHistory wh : workHistories) {
+            wh.setEmpName(wh.getEmployee().getEmpName());
+        }
+        ObjectMapper mapper = new ObjectMapper();
+        String json = mapper.writeValueAsString(workHistories);
+        return json;
+    }
+
+    @GetMapping("/ajax/getworkhistory/{workHistNo}")
+    @ResponseBody
+    public String getJsonWorkHistories(@PathVariable("workHistNo") int workHistNo, HttpSession session) throws JsonProcessingException {
+        int branchNo = (Integer) (session.getAttribute("branchno"));
+        WorkHistory workHistory = workHistoryService.getWorkHistory(workHistNo,branchNo);
+        ObjectMapper mapper = new ObjectMapper();
+        String json = mapper.writeValueAsString(workHistory);
+        return json;
+    }
+
+    @PostMapping("/ajax/smartgenworkhistory")
+    @ResponseBody
+    public String smartGenWorkHistory(HttpSession session, HttpServletRequest request, HttpServletResponse response) throws JsonProcessingException {
+        int branchNo = (Integer) (session.getAttribute("branchno"));
+        Date d = Date.valueOf(request.getParameter("date"));
+        List<WorkHistory> workHistories = workHistoryService.findByWorkDate(d);
+        List<EmployeeTable> employeeTables = employeeTableService.findAllByDate(d);
+        Set<Integer> workEmpNo = new HashSet<Integer>();
+        Set<Integer> tableEmpNo = new HashSet<Integer>();
+
+        for (WorkHistory wh : workHistories) {
+            workEmpNo.add(wh.getEmpNo());
+        }
+
+        for (EmployeeTable et : employeeTables) {
+            tableEmpNo.add(et.getEmpNo());
+        }
+        tableEmpNo.removeAll(workEmpNo);
+        ObjectMapper mapper = new ObjectMapper();
+        String json = mapper.writeValueAsString(tableEmpNo);
+        return json;
+    }
+
+    @PostMapping("/ajax/filterworkhistory")
+    @ResponseBody
+    public String filterWorkByDate(HttpSession session, HttpServletRequest request, HttpServletResponse response) throws JsonProcessingException {
+        String filterDate = request.getParameter("filterdate");
+        String fromddmmyyyy = filterDate.substring(0,10);
+        String toddmmyyyy = filterDate.substring(13,23);
+        Date fromDate = Date.valueOf(fromddmmyyyy.substring(6,10)+"-"+fromddmmyyyy.substring(3,5)+"-"+fromddmmyyyy.substring(0,2));
+        Date toDate = Date.valueOf(toddmmyyyy.substring(6,10)+"-"+toddmmyyyy.substring(3,5)+"-"+toddmmyyyy.substring(0,2));
+        List<WorkHistory> workHistories = workHistoryService.findByWorkDate(fromDate,toDate);
+        for (WorkHistory wh : workHistories) {
+            wh.setEmpName(wh.getEmployee().getEmpName());
+        }
+        ObjectMapper mapper = new ObjectMapper();
+        String json = mapper.writeValueAsString(workHistories);
+        return json;
+    }
     //----------------------------------------------------------------------------------------------------------//
 
     @RequestMapping("/pay")
@@ -168,17 +266,23 @@ public class EmployeeController {
         int branchNo = (Integer) (session.getAttribute("branchno"));
         List<Employee> employees = employeeService.findByBranchNo(branchNo);
         model.addAttribute("employees", employees);
-        return "/WEB-INF/emppaid.jsp";
+        return EMP_PATH+"/emppaid.jsp";
     }
 
-    @GetMapping("/pay/{empNo}")
-    public void pay(@PathVariable("empNo") int empNo, Model model, HttpServletRequest request, HttpServletResponse response) throws IOException {
+    @PostMapping("/payforemp")
+    public void pay(Model model, HttpServletRequest request, HttpServletResponse response, HttpSession session) throws Exception {
+        int empNo = Integer.parseInt(request.getParameter("empNo"));
+        int branchNo = (Integer) (session.getAttribute("branchno"));
         double pay = Double.parseDouble(request.getParameter("pay"));
-        EmployeePay employeePay = new EmployeePay();
-        employeePay.setPay(pay);
-        employeePay.setEmpNo(empNo);
-        employeePayService.save(employeePay);
-        response.sendRedirect(request.getContextPath() + "/employee/pay");
+        double remainPay = workHistoryService.getWorkSum(branchNo, empNo) - employeePayService.getPaySum(branchNo, empNo);
+        if (pay <= remainPay) {
+            EmployeePay employeePay = new EmployeePay();
+            employeePay.setPay(pay);
+            employeePay.setEmpNo(empNo);
+            employeePayService.save(employeePay);
+        } else {
+            throw new Exception();
+        }
     }
 
     //----------------------------------------------------------------------------------------------------------//
@@ -192,7 +296,7 @@ public class EmployeeController {
         model.addAttribute("employees", employees);
         model.addAttribute("employeePositions", employeePositions);
         model.addAttribute("employeeTables", employeeTables);
-        return "/WEB-INF/emptable.jsp";
+        return EMP_PATH+"/emptable.jsp";
     }
 
     @RequestMapping("/managetable")
@@ -201,16 +305,15 @@ public class EmployeeController {
         String[] stEmpTimeNo = request.getParameterValues("empTimeNo");
         String[] stEmpNo = request.getParameterValues("empNo");
         String[] stEmpPosNo = request.getParameterValues("empPosNo");
-        String[] stTimeStart = request.getParameterValues("timeStart");
-        String[] stTimeEnd = request.getParameterValues("timeEnd");
+        String[] stWorkTime = request.getParameterValues("workTime"); // Kopi
         List<EmployeeTable> employeeTables = new ArrayList<EmployeeTable>();
         for (int i = 0; i < stEmpNo.length; i++) {
             EmployeeTable employeeTable = new EmployeeTable();
             int empNo = Integer.parseInt(stEmpNo[i]);
             int empPosNo = Integer.parseInt(stEmpPosNo[i]);
-            Time timeStart = Time.valueOf(stTimeStart[i] + ":00");
-            Time timeEnd = Time.valueOf(stTimeEnd[i] + ":00");
-            if(stEmpTimeNo != null){
+            Time timeStart = Time.valueOf(stWorkTime[i]); // Kopi
+            Time timeEnd = Time.valueOf(stWorkTime[i]); // Kopi
+            if (stEmpTimeNo != null) {
                 employeeTable.setEmpTimeNo(Integer.parseInt(stEmpTimeNo[i]));
             }
             employeeTable.setDate(date);
@@ -227,79 +330,9 @@ public class EmployeeController {
 
     @Transactional
     @RequestMapping("/deletetable/{empTimeNo}")
-    public void removeByEmpTimeNo(@PathVariable("empTimeNo") int empTimeNo,HttpServletRequest request,HttpServletResponse response,HttpSession session) throws IOException {
-        int branchNo = (Integer)(session.getAttribute("branchno"));
+    public void removeByEmpTimeNo(@PathVariable("empTimeNo") int empTimeNo, HttpServletRequest request, HttpServletResponse response, HttpSession session) throws IOException {
+        int branchNo = (Integer) (session.getAttribute("branchno"));
         employeeTableService.removeByEmpTimeNoAndBranchNo(empTimeNo, branchNo);
         response.sendRedirect(request.getContextPath() + "/employee/table");
-    }
-
-    //----------------------------------------------------------------------------------------------------------//
-
-    @RequestMapping("/check")
-    public String toEmployeeCheck(Model model, HttpSession session) {
-        int branchNo = (Integer) (session.getAttribute("branchno"));
-        Date date = new Date(System.currentTimeMillis());
-        Time time = new Time(System.currentTimeMillis());
-        List<EmployeeTable> employeeTables = employeeTableService.findEmployeeTableNow(branchNo, date, time);
-        model.addAttribute("employeeTables", employeeTables);
-        return "/WEB-INF/empcheck.jsp";
-    }
-
-    @RequestMapping("/clockin")
-    public void clockIn(HttpServletRequest request, HttpServletResponse response) throws IOException {
-        int empNo = Integer.parseInt(request.getParameter("empNo"));
-        int empTimeNo = Integer.parseInt(request.getParameter("empTimeNo"));
-        Date date = new Date(System.currentTimeMillis());
-        Time workStart = new Time(System.currentTimeMillis());
-        WorkHistory workHistory = new WorkHistory();
-        workHistory.setWorkDate(date);
-        workHistory.setWorkStart(workStart);
-        workHistory.setEmpNo(empNo);
-        workHistory.setEmpTimeNo(empTimeNo);
-        workHistoryService.save(workHistory);
-        response.sendRedirect(request.getContextPath() + "/employee/check");
-    }
-
-    @RequestMapping("/clockout")
-    public void clockOut(HttpServletRequest request, HttpServletResponse response) throws IOException {
-        int workHistNo = Integer.parseInt(request.getParameter("workHistNo"));
-        WorkHistory workHistory = workHistoryService.findByWorkHistNo(workHistNo);
-        Employee employee = workHistory.getEmployee();
-
-        Time workStart = workHistory.getWorkStart();
-        Time workEnd = new Time(System.currentTimeMillis());
-
-        LocalTime localWorkStart = workStart.toLocalTime();
-        LocalTime localWorkEnd = workEnd.toLocalTime();
-
-        int hourStart = localWorkStart.getHour();
-        int minStart = localWorkStart.getMinute();
-        int hourEnd = localWorkEnd.getHour();
-        int minEnd = localWorkEnd.getMinute();
-
-        int hour = 0,minute = 0;
-
-        if(minEnd < minStart){
-            hour = hourEnd - hourStart - 1;
-            minute = (60 - minStart) + minEnd;
-        }else{
-            hour = hourEnd - hourStart;
-            minute = minEnd - minStart;
-        }
-
-
-        double workPay = employee.getPay();
-        if (employee.getPayType().equals(Employee.HOUR)) {
-            workPay = workPay * ( hour + ((minute*5)/300d) );
-        }
-
-        workHistory.setWorkEnd(workEnd);
-        workHistory.setWorkPay(workPay);
-        workHistory.setWorkHour(hour);
-        workHistory.setWorkMin(minute);
-
-        workHistoryService.save(workHistory);
-
-        response.sendRedirect(request.getContextPath() + "/employee/check");
     }
 }
